Organized code a little bit, added batch
Now i'm working on the loading mode- to let the game be initialized from the file. To make game more playable i want to make a game loader which will use multimethreading to load a batch, when player operates on the other one. Also if it'll be posible make code more clean, especially whole the Entity class